delete user and export documents now in user account page
Deleting a user from your account management page makes more sense.
Also, so far this action was triggered by a GET request, now it's
a more reasonable POST request, which is the suggested method for
destructive things.

Features page is now more prominently linked to from the history or
landing page. It was previously below the user dropdown.

Signed-off-by: Claudius <opensource@amenthes.de>