Rework config management
At the moment, CodiMD has an self-written config management library
which works more or less great. It leads to confusion about config
values, as well as providing semi-great feedback about the values.

The new library called node-convict is written and maintained by Mozilla
and provides schema-based validation for configuration. allows easier
extension of the config mangement to formats like yaml and commandline
arguments. And can be used to improve the configs themselves over time.

This patch is a first step towards massively improving the config
managment. Next steps should be:

- Let components like the github-login-integration provide an own schema
that is used for validation.
- Add more config file formats
- Improve formats for existing config variables
- Further reduce the handling of own config settings.
- Split actual configs from internal state knowledge (like
`isGitHubEnabled`)
- Writing docs for the new config management

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>