Move TLS config into an own subsection
In order to further restructure and modulize the configs of CodiMD,
moving TLS support into an own config section should work towards that
goal and is a first PoC on how this should be done in future.

This patch places all TLS related config under the config section `tls`
and adds a `tls.enable` variable to it, which allows to explicitly
enable or disable TLS and should be more intuitive than `useSSL`.

It should be considered whenever we want to establish a `.enable` key
for all optional module configs as we already have if to HSTS, CSP and
now TLS, but for example not for GitHub, GitLab, and other auth
providers.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>