Remove Tests for EOL node version 8
Node 8 is End of Life since the beginning of 2020.[1] Due to not
deprecating it earlier, the next release will be the last release
supporting it. There are no breaking changes to be expected anymore,
therefore removing the Tests can be considered safe and the release can
start its existence with a green CI.

This patch removes the test for NodeJS version 8 from the TravisCI jobs.

[1]: https://nodejs.org/en/about/releases/