Updated schedule of renovate (#192)
* Updated schedule of renovate

Even if renovate only creates 1-2 pull requests per day, they start to become annoying. Therefore, we decided to limit the timeslot for renovate actions to our dependency-patch-day Saturday.

* Added option to remove the hourly limit

Because we extend config:base we have a default hourly limit of 2. This change sets that to 0 and thereby disables the limit.