Add guest note autolock feature
This patch adds a basic implementation for an autolock feature for guest
notes. It uses the config options `CMD_LOCKGUESTNOTES_TIMEOUT` to set
a timeout in seconds when a note should be locked. As well as
`CMD_LOCKGUESTNOTES_PERMISSION` to set the desired permission.

It is possible to extend this feature later on to allow a deletion of
notes or to lock the note based on the absence of activity instead of
statically based on its creation date. However this patch doesn't
facilitate those possiblities.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>