Cache-bust permissions for realtime editing
This patch adjust the `ifMayEdit`-function in the realtime code, that is
responsible for checking permissions for users. Currently, since the
only way to adjust permissions is by using the Webinterface and its
web-sockets, the realtime functionality used a cached version of the
note. This should reduce stress and latency on the database, but at the
cost of validating the actual permission set in the database.

This patch ends this behaviour by cache-busting the permission check.
While the realtime code continues to use the cached note version for all
other operations, this patch forces a permission check on the database
backend for each operation.

This might have a huge performance pentality, therefore excessive
testing is mandatory to verify that this change doesn't come with a
drawback that makes larger instances unusable.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>