Rework permission logic in realtime code
This patch adjust the boolean logic used in the realtime code to check
permissions of a note during OT by reducing the amount of `if`s and
negations in the expressions. It also changes to a deny by default
and rearranges the switch-case statement a bit.

The patch originates from a comment during the code review, as the
previous commit, touches onto this code and introduces cache-busting for
the permission check. It was hinted that the current code is overly
complicated and hard to read. This patch aims to improve the code
quality of this function and make it easier to understand.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>