Docs: Include docker-compose snippet
Directing users to use the container repo to install
HedgeDoc has caused upgrade-trouble in the past, after we made
incompatible changes to the docker-compose.yml in the repo.

This commit inserts a minimal docker-compose snipped in the
docs, which can be used to get started. Upgrading can then
be performed by just incrementing the referenced image version.

Signed-off-by: David Mehren <git@herrmehren.de>